initial combined suggestion and comment working
