+++ packages/tinacms/src/toolkit/fields/plugins/discussion-plugin/index.ts
@@ -3,3 +3,4 @@ export * from './comment-plugin';
 export * from './comment-state-synchronizer';
 export * from './comment-toolbar-button';
 export * from './comment-node';
+export * from './annotation-popover';
