+++ packages/tinacms/src/toolkit/fields/plugins/mdx-field-plugin/plate/index.tsx
@@ -12,9 +12,8 @@ import { useCreateEditor } from './hooks/use-create-editor';
 import { editorPlugins } from './plugins/editor-plugins';
 import { FloatingToolbar } from './components/plate-ui/floating-toolbar';
 import FloatingToolbarButtons from './components/floating-toolbar-buttons';
-import { CommentPopover } from '../../discussion-plugin/comment-popover';
+import { AnnotationPopover } from '../../discussion-plugin/annotation-popover';
 import { CommentStateSynchronizer } from '../../discussion-plugin/comment-state-synchronizer';
-import { SuggestionPopover } from '../../suggestion-plugin';
 
 export const RichEditor = ({ input, tinaForm, field }: RichTextType) => {
   const initialValue = React.useMemo(() => {
@@ -89,8 +88,7 @@ export const RichEditor = ({ input, tinaForm, field }: RichTextType) => {
                   <FloatingToolbarButtons />
                 </FloatingToolbar>
               ) : null}
-              <CommentPopover />
-              <SuggestionPopover />
+              <AnnotationPopover />
             </ToolbarProvider>
             <Editor />
           </TooltipProvider>
